added quick and dirty commandline interface
