+++ .gitignore
@@ -1,0 +1,4 @@
+.vs/
+**/bin
+**/obj
+**/packages
