+++ VBEThemeColorEditor/VBEThemeColorEditorCli/packages.config
@@ -1,0 +1,4 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="CommandLineParser" version="2.8.0" targetFramework="net452" />
+</packages>
